requirements.txt
fastapi~=0.75.0
uvicorn==0.17.6
numpy>=1.26.0  
pydantic==1.9.1
Pillow==9.1.1
tensorflow==2.18.0
python-multipart
